# Set 'mode: :compat' in Oj.dump to stringify keys