All in the title

### Performance
this change slows down the performance a little


#### Ruby 2.7
with `mode: :compat`
```
  JSON short content     93.920k (± 0.7%) i/s -    478.533k in   5.095383s
Logger short content    106.209k (± 0.8%) i/s -    536.100k in   5.047894s
   JSON long content     45.085k (±11.5%) i/s -    227.424k in   5.142470s
 Logger long content     35.762k (±34.6%) i/s -    156.832k in   5.043390s
```

without `mode: :compat`
```
  JSON short content    107.124k (± 6.2%) i/s -    545.670k in   5.112505s
Logger short content    109.938k (± 1.1%) i/s -    555.500k in   5.053470s
   JSON long content     45.739k (±13.6%) i/s -    223.605k in   5.019013s
 Logger long content     34.799k (±30.5%) i/s -    158.652k in   5.006122s
```
#### Ruby 3.0
with `mode: :compat`
```
  JSON short content     88.108k (± 3.4%) i/s -    441.490k in   5.016234s
Logger short content     97.983k (± 3.7%) i/s -    493.773k in   5.046516s
   JSON long content     42.026k (± 3.8%) i/s -    212.670k in   5.068215s
 Logger long content     31.937k (±14.6%) i/s -    161.107k in   5.135188s
```

without `mode: :compat`
```
  JSON short content     98.295k (± 1.5%) i/s -    499.647k in   5.084235s
Logger short content     97.783k (± 2.5%) i/s -    491.088k in   5.025475s
   JSON long content     44.083k (± 6.8%) i/s -    224.077k in   5.116877s
 Logger long content     30.401k (± 8.0%) i/s -    155.190k in   5.138547s
```
#### Ruby 3.1
with `mode: :compat`
```
  JSON short content     89.399k (± 0.6%) i/s -    449.150k in   5.024249s
Logger short content     98.255k (± 1.8%) i/s -    494.450k in   5.034118s
   JSON long content     43.882k (± 6.2%) i/s -    223.150k in   5.109795s
 Logger long content     30.615k (± 7.8%) i/s -    152.199k in   5.001883s
```

without `mode: :compat`
```
  JSON short content     96.925k (± 0.7%) i/s -    488.400k in   5.039164s
Logger short content     98.041k (± 7.3%) i/s -    492.650k in   5.075970s
   JSON long content     44.712k (±11.3%) i/s -    226.368k in   5.153677s
 Logger long content     31.410k (± 9.0%) i/s -    156.828k in   5.033354s
```

@jcarres-mdsol @jfeltesse-mdsol