@ykitamura-mdsol 2022-04-29
slight improvement by enabling caching: a6a7b9a2afde916e8767346c0ac6793e81c0a3d2

#### Ruby 2.7
```
  JSON short content    107.294k (± 1.2%) i/s -    545.116k in   5.081370s
Logger short content    127.683k (± 1.5%) i/s -    650.600k in   5.096536s
   JSON long content     47.747k (±18.2%) i/s -    231.836k in   5.079736s
 Logger long content     32.495k (±12.5%) i/s -    164.970k in   5.162471s
```

#### Ruby 3.0
```
  JSON short content     96.013k (± 0.8%) i/s -    485.650k in   5.058447s
Logger short content    109.737k (± 0.9%) i/s -    553.962k in   5.048498s
   JSON long content     44.041k (± 8.4%) i/s -    224.518k in   5.146373s
 Logger long content     30.207k (± 8.3%) i/s -    154.575k in   5.153832s
```

#### Ruby 3.1
```
  JSON short content     89.546k (± 0.8%) i/s -    454.359k in   5.074395s
Logger short content     98.843k (± 0.8%) i/s -    504.100k in   5.100321s
   JSON long content     44.100k (±11.3%) i/s -    215.401k in   5.100501s
 Logger long content     31.313k (±16.7%) i/s -    159.572k in   5.187550s
```